odradjen deo za delete

+++ src/main/java/skprojekat/flightservice/service/impl/PlaneServiceImpl.java
@@ -7,18 +7,24 @@ import org.springframework.stereotype.Service;
 import skprojekat.flightservice.dto.PlaneCreateDto;
 import skprojekat.flightservice.dto.PlaneDto;
 import skprojekat.flightservice.mapper.PlaneMapper;
+import skprojekat.flightservice.model.Flight;
+import skprojekat.flightservice.repository.FlightRepository;
 import skprojekat.flightservice.repository.PlaneRepository;
 import skprojekat.flightservice.service.PlaneService;
+
+import java.util.Optional;
 
 @Service
 public class PlaneServiceImpl implements PlaneService{
 
 	private PlaneRepository planeRepo;
 	private PlaneMapper planeMapper;
+	private FlightRepository flightRepo;
 	
-	public PlaneServiceImpl(PlaneRepository planeRepo, PlaneMapper planeMapper) {
+	public PlaneServiceImpl(PlaneRepository planeRepo, PlaneMapper planeMapper, FlightRepository flightRepo) {
 		this.planeMapper = planeMapper;
 		this.planeRepo = planeRepo;
+		this.flightRepo = flightRepo;
 	}
 	
 	@Override
@@ -42,6 +48,9 @@ public class PlaneServiceImpl implements PlaneService{
 
 	@Override
 	public void deleteById(Integer id) {
+		Optional<Flight> flights = flightRepo.findByPlane_Id(id);
+		if(flights.isPresent())
+			return;
 		planeRepo.deleteById(id);
 	}
